refactor: 💡 png 이미지 형식을 webp로 변경
